+++ modules/documentation/docs/changelog.md
@@ -1,5 +1,46 @@
 # Vector Changelog
 
+## Next Release
+
+## 0.1.15-beta.1
+
+- \[router\] Do not forward if deposit balance is not present in channel
+- \[router\] Fix start script to pass mnemonic secret into router
+## 0.1.15-beta.0
+
+- \[utils\] Don't use `instanceof` operator
+
+## 0.1.14
+
+- \[server-node\] Fix withdrawal commitment store bugs
+- \[server-node\] Fix error handling (always use `jsonifyError`)
+
+## 0.1.13
+
+- \[router\] Hotfix for gas price query errors
+- \[router\] Admin endpoint to retry withdrawal txs
+
+## 0.1.13-beta.0
+
+- \[all\] Update packages that were flagged by `npm audit`
+
+## 0.1.12
+
+- \[all\] Update packages that were flagged by `npm audit`
+
+## 0.1.11
+
+- \[all\] Update packages that were flagged by `npm audit`
+
+## 0.1.10
+
+- \[utils\] Pass in `chainId` when creating providers
+
+## 0.1.9
+
+- \[engine\] Fix deposit reconciliation race condition evaluation
+- \[contracts\] Use `gasNow` to get mainnet gas price over ethers
+
 ## 0.1.8
 
 - \[node\] Add `getTransfers` endpoint with filtering capabilities.